[Refactor] Default Layout 컴포넌트 합침
