+++ src/components/layouts/DefaultLayout.tsx
@@ -1,18 +1,30 @@
 import React, { useEffect } from 'react';
 import { useRouter } from 'next/router';
-import Header from './Header';
-import SideBar from './SideBar';
+import { logout } from '@/utils/login';
 import styles from '@/styles/layout.module.scss';
+import sideStyles from '@/styles/sidebar.module.scss';
+import headerStyles from '@/styles/header.module.scss';
 import { AccessType } from '@/types/accessType';
 import { useUser } from '@/utils/hooks/useUser';
 import { toast } from 'react-hot-toast';
+import navigationData from '@/data/navigationData.json';
+import Link from 'next/link';
+
+type navigationItem = {
+  path: string;
+  name: string;
+  accessLevel: AccessType;
+};
+
+type navigationData = navigationItem[];
 
 export default function DefaultLayout(props: {
   children: React.ReactNode;
   auth: AccessType;
 }) {
   const router = useRouter();
-  const { user, loading, error } = useUser();
+  const currentPath = router.pathname;
+  const { user, loading, error, resetUser } = useUser();
 
   useEffect(() => {
     if (error) {
@@ -28,7 +40,7 @@ export default function DefaultLayout(props: {
     return null;
   }
 
-  if (loading) {
+  if (loading || !user) {
     return <div>loading...</div>;
   }
 
@@ -37,10 +49,40 @@ export default function DefaultLayout(props: {
     return null;
   }
 
+  const handleLogout = async () => {
+    await logout();
+    resetUser();
+    router.push('/login');
+  };
+
   return (
     <div className={styles.generalWrapper}>
-      <Header />
-      <SideBar />
+      <header className={headerStyles.wrapper}>
+        <div className={headerStyles.title}>Routing Practice</div>
+        <button onClick={handleLogout} className={headerStyles.button}>
+          Logout
+        </button>
+      </header>
+      <nav className={sideStyles.wrapper}>
+        <ul className={sideStyles.ulWrapper}>
+          {navigationData
+            .filter((item) => item.accessLevel <= user.role)
+            .map((item) => (
+              <li key={item.name}>
+                <Link
+                  className={
+                    item.path === currentPath
+                      ? sideStyles.currentLink
+                      : sideStyles.link
+                  }
+                  href={item.path}
+                >
+                  {item.name}
+                </Link>
+              </li>
+            ))}
+        </ul>
+      </nav>
       <div className={styles.bodyWrapper}>{props.children}</div>
     </div>
   );
